# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/security/mfa/setup": {
      "post": {
        "operationId": "AuthSecurityController_setupMFA",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MFASetupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MFA设置生成成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "设置MFA",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/mfa/verify": {
      "post": {
        "operationId": "AuthSecurityController_verifyMFA",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MFAVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MFA验证成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "验证MFA代码并启用",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/mfa/disable": {
      "post": {
        "operationId": "AuthSecurityController_disableMFA",
        "parameters": [],
        "responses": {
          "200": {
            "description": "MFA禁用成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "禁用MFA",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/mfa/status": {
      "get": {
        "operationId": "AuthSecurityController_getMFAStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "MFA状态获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取MFA状态",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/mfa/backup-codes/regenerate": {
      "post": {
        "operationId": "AuthSecurityController_regenerateBackupCodes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "备用代码重新生成成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "重新生成备用代码",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/devices": {
      "get": {
        "operationId": "AuthSecurityController_getUserDevices",
        "parameters": [],
        "responses": {
          "200": {
            "description": "设备列表获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取用户设备列表",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/devices/{deviceId}/trust": {
      "post": {
        "operationId": "AuthSecurityController_trustDevice",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "设备信任成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "信任设备",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/devices/{deviceId}/untrust": {
      "post": {
        "operationId": "AuthSecurityController_untrustDevice",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "设备取消信任成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "取消信任设备",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/devices/{deviceId}/delete": {
      "post": {
        "operationId": "AuthSecurityController_deleteDevice",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "设备删除成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除设备",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/devices/trust-settings": {
      "get": {
        "operationId": "AuthSecurityController_getTrustSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "设备信任设置获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取设备信任设置",
        "tags": [
          "Auth Security"
        ]
      },
      "post": {
        "operationId": "AuthSecurityController_setTrustSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "设备信任设置更新成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "设置设备信任设置",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/login-history": {
      "get": {
        "operationId": "AuthSecurityController_getLoginHistory",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "登录历史获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取登录历史",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/password/check": {
      "post": {
        "operationId": "AuthSecurityController_checkPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordCheckDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密码强度检查完成"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "检查密码强度",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/password/policy": {
      "get": {
        "operationId": "AuthSecurityController_getPasswordPolicy",
        "parameters": [],
        "responses": {
          "200": {
            "description": "密码策略获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取密码策略",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/password/generate": {
      "post": {
        "operationId": "AuthSecurityController_generatePassword",
        "parameters": [
          {
            "name": "length",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "安全密码生成成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "生成安全密码",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/alerts": {
      "get": {
        "operationId": "AuthSecurityController_getSecurityAlerts",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "includeResolved",
            "required": true,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "安全告警获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取安全告警",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/alerts/{alertId}/read": {
      "post": {
        "operationId": "AuthSecurityController_markAlertAsRead",
        "parameters": [
          {
            "name": "alertId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "告警标记为已读成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "标记告警为已读",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/alerts/{alertId}/resolve": {
      "post": {
        "operationId": "AuthSecurityController_resolveAlert",
        "parameters": [
          {
            "name": "alertId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "告警解决成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "解决告警",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/alerts/settings": {
      "get": {
        "operationId": "AuthSecurityController_getAlertSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "告警设置获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取告警设置",
        "tags": [
          "Auth Security"
        ]
      },
      "post": {
        "operationId": "AuthSecurityController_setAlertSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "告警设置更新成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "设置告警设置",
        "tags": [
          "Auth Security"
        ]
      }
    },
    "/auth/security/stats": {
      "get": {
        "operationId": "AuthSecurityController_getSecurityStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "安全统计信息获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取安全统计信息",
        "tags": [
          "Auth Security"
        ]
      }
    },
...
  },
  "info": {
    "title": "jmni-server",
    "description": "JMNI Server API 文档 - 供外部团队参考",
    "version": "0.0.1",
    "contact": {
      "name": "JMNI Team",
      "url": "https://jmni.cn",
      "email": "support@jmni.cn"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "https://api.jmni.cn",
      "description": "生产环境"
    },
    {
      "url": "http://localhost:2233",
      "description": "开发环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT Access Token"
      }
    },
    "schemas": {
...
      "MFASetupDto": {
        "type": "object",
        "properties": {}
      },
      "MFAVerifyDto": {
        "type": "object",
        "properties": {}
      },
      "PasswordCheckDto": {
        "type": "object",
        "properties": {}
      },
      "AlertSettingsDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}